+++ src/main/java/com/saintdan/framework/component/ValidateHelper.java
@@ -1,12 +1,17 @@
 package com.saintdan.framework.component;
 
+import com.saintdan.framework.annotation.NotNullField;
+import com.saintdan.framework.constant.ControllerConstant;
 import com.saintdan.framework.domain.ClientDomain;
 import com.saintdan.framework.enums.ErrorType;
+import com.saintdan.framework.enums.OperationType;
 import com.saintdan.framework.param.BaseParam;
 import com.saintdan.framework.param.ClientParam;
 import com.saintdan.framework.po.User;
 import com.saintdan.framework.tools.SpringSecurityUtils;
 import com.saintdan.framework.vo.ResultVO;
+import java.lang.reflect.Field;
+import org.apache.commons.lang3.ArrayUtils;
 import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -28,67 +33,59 @@ import org.springframework.validation.BindingResult;
   /**
    * Validate current user, param and sign.
    *
-   * @param result      bind result
-   * @param currentUser current user
-   * @param logger      log
+   * @param param         param bean
+   * @param result        bind result
+   * @param currentUser   current user
+   * @param logger        log
+   * @param operationType {@link OperationType}
    * @return result vo
    * @throws Exception
    */
-  public ResultVO validateWithOutSignCheck(BindingResult result, User currentUser, Logger logger) throws Exception {
+  public ResultVO validate(BaseParam param, BindingResult result, User currentUser, Logger logger, OperationType operationType) throws Exception {
     if (result.hasErrors()) {
       return resultHelper.infoResp(ErrorType.SYS0002, result.toString());
     }
-    return validateWithOutSignCheck(currentUser, logger);
+    return validate(param, currentUser, logger, operationType);
   }
 
   /**
    * Validate current user and sign.
    *
-   * @param currentUser currentUser
-   * @param logger      log
+   * @param param         param bean
+   * @param currentUser   currentUser
+   * @param logger        log
+   * @param operationType {@link OperationType}
    * @return result VO
    * @throws Exception
    */
-  public ResultVO validateWithOutSignCheck(User currentUser, Logger logger) throws Exception {
+  public ResultVO validate(BaseParam param, User currentUser, Logger logger, OperationType operationType) throws Exception {
     //check currentUser
     if (currentUser == null || currentUser.getId() == null) {
       return resultHelper.infoResp(logger, ErrorType.SYS0003, ErrorType.SYS0003.description());
+    }
+    return validate(param, operationType);
+  }
+
+  /**
+   * Bean properties null validation.
+   *
+   * @param param         Param bean
+   * @param operationType {@link OperationType}
+   * @return
+   * @throws Exception
+   */
+  public ResultVO validate(BaseParam param, OperationType operationType) throws Exception {
+    Field[] fields = param.getClass().getDeclaredFields();
+    for (Field field : fields) {
+      if (field == null || !field.isAnnotationPresent(NotNullField.class)) {
+        continue; // Ignore field without ParamField annotation.
+      }
+      field.setAccessible(true);
+      if (ArrayUtils.contains(field.getAnnotation(NotNullField.class).value(), operationType) && field.get(param) == null) {
+        return resultHelper.infoResp(ErrorType.SYS0002, String.format(ControllerConstant.PARAM_BLANK, field.getName()));
+      }
     }
     return null;
-  }
-
-  /**
-   * Validate current user, param and sign.
-   *
-   * @param result      bind result
-   * @param currentUser current user
-   * @param param       param
-   * @param logger      log
-   * @return result vo
-   * @throws Exception
-   */
-  public ResultVO validateWithSignCheck(BindingResult result, User currentUser, BaseParam param, Logger logger) throws Exception {
-    if (result.hasErrors()) {
-      return resultHelper.infoResp(ErrorType.SYS0002, result.toString());
-    }
-    return validateWithSignCheck(currentUser, param, logger);
-  }
-
-  /**
-   * Validate current user and sign.
-   *
-   * @param currentUser currentUser
-   * @param param       param
-   * @param logger      log
-   * @return result VO
-   * @throws Exception
-   */
-  public ResultVO validateWithSignCheck(User currentUser, BaseParam param, Logger logger) throws Exception {
-    //check currentUser
-    if (currentUser == null || currentUser.getId() == null) {
-      return resultHelper.infoResp(logger, ErrorType.SYS0003, ErrorType.SYS0003.description());
-    }
-    return validateWithSignCheck(param, logger);
   }
 
   /**
@@ -100,7 +97,7 @@ import org.springframework.validation.BindingResult;
    * @throws Exception
    */
 
-  public ResultVO validateWithSignCheck(BaseParam param, Logger logger) throws Exception {
+  public ResultVO validateWithRSASignCheck(BaseParam param, Logger logger) throws Exception {
     // Get current clientId
     String clientId = SpringSecurityUtils.getCurrentClientId();
     // Sign verification.