+++ src/main/java/com/saintdan/framework/controller/GroupController.java
@@ -10,6 +10,7 @@ import com.saintdan.framework.constant.VersionConstant;
 import com.saintdan.framework.domain.GroupDomain;
 import com.saintdan.framework.enums.ErrorType;
 import com.saintdan.framework.enums.OperationStatus;
+import com.saintdan.framework.enums.OperationType;
 import com.saintdan.framework.exception.CommonsException;
 import com.saintdan.framework.param.GroupParam;
 import com.saintdan.framework.po.Group;
@@ -59,7 +60,7 @@ public class GroupController {
   public ResultVO create(@CurrentUser User currentUser, @Valid GroupParam param, BindingResult result) {
     try {
       // Validate current user, param and sign.
-      ResultVO resultVO = validateHelper.validateWithOutSignCheck(result, currentUser, logger);
+      ResultVO resultVO = validateHelper.validate(param, result, currentUser, logger, OperationType.CREATE);
       if (resultVO != null) {
         return resultVO;
       }
@@ -113,7 +114,7 @@ public class GroupController {
       if (StringUtils.isBlank(id)) {
         return resultHelper.infoResp(ErrorType.SYS0002, String.format(ControllerConstant.PARAM_BLANK, ControllerConstant.ID_PARAM));
       }
-      return resultHelper.successResp(groupDomain.getById(new GroupParam(Long.valueOf(id)), GroupVO.class));
+      return resultHelper.successResp(groupDomain.getById((Long.valueOf(id)), GroupVO.class));
     } catch (CommonsException e) {
       // Return error information and log the exception.
       return resultHelper.infoResp(logger, e.getErrorType(), e.getMessage());
@@ -133,15 +134,12 @@ public class GroupController {
   @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
   public ResultVO update(@CurrentUser User currentUser, @PathVariable String id, @Valid GroupParam param, BindingResult result) {
     try {
-      if (StringUtils.isBlank(id)) {
-        return resultHelper.infoResp(ErrorType.SYS0002, String.format(ControllerConstant.PARAM_BLANK, ControllerConstant.ID_PARAM));
-      }
+      param.setId(StringUtils.isBlank(id) ? null : Long.valueOf(id));
       // Validate current user, param and sign.
-      ResultVO resultVO = validateHelper.validateWithOutSignCheck(result, currentUser, logger);
+      ResultVO resultVO = validateHelper.validate(param, result, currentUser, logger, OperationType.UPDATE);
       if (resultVO != null) {
         return resultVO;
       }
-      param.setId(Long.valueOf(id));
       // Update group.
       return resultHelper.successResp(groupDomain.update(param, currentUser));
     } catch (CommonsException e) {
@@ -162,17 +160,16 @@ public class GroupController {
   @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
   public ResultVO delete(@CurrentUser User currentUser, @PathVariable String id) {
     try {
-      if (StringUtils.isBlank(id)) {
-        return resultHelper.infoResp(ErrorType.SYS0002, String.format(ControllerConstant.PARAM_BLANK, ControllerConstant.ID_PARAM));
-      }
-      ResultVO resultVO = validateHelper.validateWithOutSignCheck(currentUser, logger);
+      GroupParam param = new GroupParam(StringUtils.isBlank(id) ? null : Long.valueOf(id));
+      // Validate current user and param.
+      ResultVO resultVO = validateHelper.validate(param, currentUser, logger, OperationType.DELETE);
       if (resultVO != null) {
         return resultVO;
       }
       // Delete group.
-      groupDomain.delete(Long.valueOf(id), currentUser);
-      final String ROLE = "group";
-      return new ResultVO(ResultConstant.OK, OperationStatus.SUCCESS, String.format(ControllerConstant.DELETE, ROLE));
+      groupDomain.delete(param, currentUser);
+      final String GROUP = "group";
+      return new ResultVO(ResultConstant.OK, OperationStatus.SUCCESS, String.format(ControllerConstant.DELETE, GROUP));
     } catch (CommonsException e) {
       // Return error information and log the exception.
       return resultHelper.infoResp(logger, e.getErrorType(), e.getMessage());